+++ examples/multitenancy/src/mailers/auth/magic_link/html.t
@@ -1,8 +1,0 @@
-;<html>
-<body>
-<p>Magic link example:</p>
-<a href="{{host}}/api/auth/magic-link/{{token}}">
-Verify Your Account
-</a>
-</body>
-</html>