+++ examples/multitenancy/src/mailers/auth/magic_link/subject.t
@@ -1,1 +1,0 @@
-Magic link example