+++ examples/multitenancy/src/mailers/auth/magic_link/text.t
@@ -1,2 +1,0 @@
-Magic link with this link:
-{{host}}/api/auth/magic-link/{{token}}