+++ examples/multitenancy/tests/models/mod.rs
@@ -1,8 +1,0 @@
-mod tenants;
-mod users;
-
-mod invoices;
-
-mod clients;
-
-mod projects;