+++ examples/multitenancy/tests/requests/mod.rs
@@ -1,9 +1,0 @@
-mod auth;
-mod frontend;
-mod prepare_data;
-
-mod clients;
-mod dashboard;
-pub mod documents;
-mod invoices;
-mod projects;