+++ examples/multitenancy/tests/tasks/mod.rs
@@ -1,2 +1,0 @@
-pub mod user_create;
-pub mod user_delete;