+++ src/model/mod.rs
@@ -3,10 +3,12 @@
 //! Useful when using `sea_orm` and want to propagate errors
 
 pub mod query;
+#[cfg(feature = "multi-tenancy")]
 mod tenant;
 use async_trait::async_trait;
 use sea_orm::DatabaseConnection;
 
+#[cfg(feature = "multi-tenancy")]
 pub use tenant::{TenantActiveModelExt, TenantEntity, TenantQueryExt};
 
 use crate::validation::ModelValidationErrors;
@@ -21,6 +23,7 @@ pub enum ModelError {
     #[error("Entity not found")]
     EntityNotFound,
 
+    #[cfg(feature = "multi-tenancy")]
     #[error("Cannot change a model's tenant")]
     TenantMismatch,