+++ src/prelude.rs
@@ -27,10 +27,9 @@ pub use crate::controller::extractor::{
     validate::{JsonValidate, JsonValidateWithMessage},
 };
 #[cfg(feature = "with-db")]
-pub use crate::model::{
-    query, Authenticable, ModelError, ModelResult, TenantActiveModelExt, TenantEntity,
-    TenantQueryExt,
-};
+pub use crate::model::{query, Authenticable, ModelError, ModelResult};
+#[cfg(feature = "multi-tenancy")]
+pub use crate::model::{TenantActiveModelExt, TenantEntity, TenantQueryExt};
 pub use crate::{
     app::{AppContext, Initializer},
     bgworker::{BackgroundWorker, Queue},